requirements.txt
spotipy==2.19.0
python-dotenv==0.19.2
Flask==2.0.1
Werkzeug==2.0.1
